persistence

+++ implementation/tourney-domain/src/main/java/com/pduda/tourney/domain/repository/TourneyRepo.java
@@ -2,6 +2,6 @@ package com.pduda.tourney.domain.repository;
 
 import com.pduda.tourney.domain.Tourney;
 
-public interface TourneyRepo extends BaseRepo<Tourney, Integer> {
+public interface TourneyRepo extends BaseRepo<Tourney, Long> {
 
 }
